+++ R/googleAuthR_shiny.R
@@ -252,6 +252,7 @@ gar_shiny_getToken <- function(code,
 #'     ))
 #' }
 reactiveAccessToken <- function(session){
+  message("reactiveAccessToken is deprecated as of googleAuthR 0.3.0. Use googleAuth() and googleAuthUI() instead.")
   shiny::reactive({
     ## gets all the parameters in the URL. The auth code should be one of them.
     
@@ -347,6 +348,7 @@ reactiveAccessToken <- function(session){
 #'     ))
 #' }
 loginOutput <- function(output_name){
+  message("loginOutput is deprecated as of googleAuthR 0.3.0. Use googleAuth() and googleAuthUI() instead.")
   shiny::uiOutput(output_name)
 }
 
@@ -450,7 +452,7 @@ renderLogin <- function(session,
                         access_type = c("online","offline"),
                         approval_prompt = c("auto","force"),
                         revoke = FALSE){
-  
+  message("renderLogin is deprecated as of googleAuthR 0.3.0. Use googleAuth() and googleAuthUI() instead.")
   access_type <- match.arg(access_type)
   approval_prompt <- match.arg(approval_prompt)
   
@@ -482,6 +484,174 @@ renderLogin <- function(session,
       
     }
   })
+}
+
+#' A Login button (Shiny Module)
+#' 
+#' UI part of shiny module, use with \link{googleAuth}
+#' 
+#' @param id shiny id
+#' 
+#' @return A shiny UI for logging in
+#' 
+#' @family shiny module functions
+#' @export
+googleAuthUI <- function(id){
+  
+  ns <- shiny::NS(id)
+  
+  shiny::uiOutput(ns("googleAuthUi"))
+}
+  
+  
+#' Server side google auth (Shiny Module)
+#' 
+#' Server part of shiny module, use with \link{googleAuthUI}
+#' 
+#' Call via \code{shiny::callModule(googleAuth, "your_ui_name", login_text = "Login")}
+#'
+#' @param login_text What the login text will read on the button
+#' @param logout_text What the logout text will read on the button
+#' @param login_class The CSS class for the login link
+#' @param logout_class The CSS class for the logout link
+#' @param access_type Online or offline access for the authentication URL
+#' @param approval_prompt Whether to show the consent screen on authentication
+#' @param revoke If TRUE a user on logout will need to re-authenticate
+#' 
+#' @return A reactive authentication token
+#' 
+#' @examples 
+#' 
+#' \dontrun{
+#' options("googleAuthR.scopes.selected" = 
+#'   c("https://www.googleapis.com/auth/urlshortener"))
+#' 
+#' shorten_url <- function(url){
+#'   body = list(
+#'     longUrl = url
+#'  )
+#'  
+#'  f <- 
+#'    gar_api_generator("https://www.googleapis.com/urlshortener/v1/url",
+#'                      "POST",
+#'                      data_parse_function = function(x) x$id)
+#'                         
+#'  f(the_body = body)
+#'  
+#'  }
+#' 
+#' server <- function(input, output, session){
+#' 
+#'   ## Create access token and render login button
+#'   access_token <- callModule(googleAuth, 
+#'                              "loginButton",
+#'                              login_text = "Login1")
+#' 
+#'   short_url_output <- eventReactive(input$submit, {
+#'     ## wrap existing function with_shiny
+#'     ## pass the reactive token in shiny_access_token
+#'     ## pass other named arguments
+#'     with_shiny(f = shorten_url, 
+#'                shiny_access_token = access_token(),
+#'                url=input$url)
+#'   })
+#' 
+#'   output$short_url <- renderText({
+#' 
+#'     short_url_output()
+#' 
+#'   })
+#' 
+#' }
+#' 
+#' ## ui
+#' ui <- fluidPage(
+#'   googleAuthUI("loginButton"),
+#'   textInput("url", "Enter URL"),
+#'   actionButton("submit", "Shorten URL"),
+#'   textOutput("short_url")
+#' )
+#' 
+#' shinyApp(ui = ui, server = server)
+#' }
+#' 
+#' @family shiny module functions
+#' @export
+googleAuth <- function(input, output, session, 
+                       login_text="Login via Google",
+                       logout_text="Logout",
+                       login_class="btn btn-primary",
+                       logout_class="btn btn-default",
+                       access_type = c("online","offline"),
+                       approval_prompt = c("auto","force"),
+                       revoke = FALSE){
+  
+  access_type <- match.arg(access_type)
+  approval_prompt <- match.arg(approval_prompt)
+  ns <- session$ns
+  
+  accessToken <- shiny::reactive({
+    
+      ## gets all the parameters in the URL. The auth code should be one of them.
+      if(!is.null(authReturnCode(session))){
+        ## extract the authorization token
+        app_url <- gar_shiny_getUrl(session)    
+        access_token <- gar_shiny_getToken(authReturnCode(session), app_url)
+        
+        Authentication$set("public", "app_url", app_url, overwrite=TRUE)
+        Authentication$set("public", "shiny", TRUE, overwrite=TRUE)
+        
+        access_token
+        
+      } else {
+        NULL
+      }
+    })
+  
+  output$googleAuthUi <- shiny::renderUI({
+    
+
+    
+    if(is.null(shiny::isolate(accessToken()))) {
+      shiny::actionLink(ns("signed_in"),
+                        shiny::a(login_text, 
+                                 href = gar_shiny_getAuthUrl(gar_shiny_getUrl(session), 
+                                                             access_type = access_type,
+                                                             approval_prompt = approval_prompt), 
+                                 class=login_class, 
+                                 role="button"))
+    } else {
+      if(revoke){
+        
+        logout_button <- shiny::actionButton(ns("revoke"), "Revoke Access", 
+                                             href = gar_shiny_getUrl(session), 
+                                             class=logout_class,
+                                             role="button")
+        
+      } else {
+        logout_button <- shiny::a(logout_text, 
+                                  href = gar_shiny_getUrl(session), 
+                                  class=logout_class, 
+                                  role="button")
+      }
+      
+      logout_button
+      
+    }
+  })
+
+  shiny::observeEvent(input[[ns("revoke")]], {
+      
+    ## GETS the revoke URL for this user's access_token
+    httr::GET(httr::modify_url("https://accounts.google.com/o/oauth2/revoke",
+                               query = 
+                                 list(token = 
+                                        shiny::isolate(access_token)$credentials$access_token)))
+    myMessage("Revoked access", level=2)
+  })
+
+  return(accessToken)
+
 }
 
 #' Listens for a user revoking authentication
@@ -566,7 +736,7 @@ renderLogin <- function(session,
 #'     ))
 #' } 
 revokeEventObserver <- function(access_token, input){
-  
+  message("revokeEventObserver is deprecated as of googleAuthR 0.3.0. Use googleAuth() and googleAuthUI() instead.")  
   shiny::observeEvent(input$revoke, {
 
         ## GETS the revoke URL for this user's access_token